Set visibility to hidden when not active
When rules modal was not active it was only set to opacity: 0. This meant that the game buttons would still be covered my the modal and could not be clicked.

Changed the css class Rules_hidden so that the modal was set to visibility: hidden. Buttons can now be clicked.